requirements.txt
absl-py==2.0.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.1.0
cachecontrol==0.13.1
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==4.2.5
django-allauth==0.58.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-webpush==0.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
firebase-admin==6.2.0
flatbuffers==23.5.26
gast==0.4.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.58.0
grpcio-status==1.59.3
h11==0.14.0
h5py==3.9.0
http-ece==1.1.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
img2pdf==0.5.0
importlib-metadata==6.8.0
incremental==22.10.0
inexactsearch==1.0.2
joblib==1.3.2
keras==2.13.1
kombu==5.3.2
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
msgpack==1.0.7
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.3.4
opt-einsum==3.3.0
packaging==23.1
pdf-mail==3.0.0
pika==1.3.2
pikepdf==8.7.1
Pillow==10.0.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.9
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic-core==2.14.3
pyfcm==1.5.4
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywebpush==1.9.4
qrcode==7.4.2
redis==5.0.1
regex==2023.8.8
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
silpa-common==0.3
six==1.16.0
sniffio==1.3.0
soundex==1.1.3
spellchecker==0.4
sqlparse==0.4.4
stripe==7.6.0
termcolor==2.3.0
tqdm==4.66.1
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.6
werkzeug==2.3.7
wrapt==1.15.0
zipp==3.17.0
zope.interface==6.0
